Compile slurm for the os of the submitter
So that it can submit jobs to a cluster that uses a different OS.

Automate the configuration of the cluster manager and submitters in RES.

Fix errors in configuration scripts.

Add a parameter to the ParallelCluster custom resource that has a hash
of all of the config file contents so that the resource gets updated
whenever anything changes.
This current issue with this implementation is that the ParallelCluster
still doesn't change so even though the custom resource is updated,
the cluster doesn't because it's configuration is unchanged.

Fix a bug in configuring licenses.
After initial configuration the playbook was failing on subsequent updates.

Support custom munge key which must be specified in secretsmanager

Try to allow RES /home to be mounted, but currently causes a validation error.

Update EDA instance types to the latest and so that all memory sizes get built.

Restrict lambda function IAM roles and cleanup cfn_nag errors and warnings.

Resolves #177
Resolves #178
Resolves #180
Resolves #181

+++ source/resources/parallel-cluster/config/bin/submitter_deconfigure.sh
@@ -1,4 +1,6 @@
 #!/bin/bash -xe
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: MIT-0
 
 full_script=$(realpath $0)
 script_dir=$(dirname $full_script)
@@ -7,14 +9,20 @@ base_script=$(basename $full_script)
 date
 echo "Started submitter_deconfigure.sh: $full_script"
 
+ErrorSnsTopicArn={{ErrorSnsTopicArn}}
+
 config_dir={{SubmitterSlurmConfigDir}}
 config_bin_dir=$config_dir/bin
 
 temp_config_dir=/tmp/{{ClusterName}}_config
-rm -rf $temp_config_dir
-cp -r $config_dir $temp_config_dir
+temp_config_bin_dir=$temp_config_dir/bin
+if [[ $script_dir != $temp_config_bin_dir ]]; then
+    rm -rf $temp_config_dir
+    cp -r $config_dir $temp_config_dir
+    exec $temp_config_dir/bin/$base_script
+fi
 
-# Configure using ansible
+# Install ansible
 if ! yum list installed ansible &> /dev/null; then
     yum install -y ansible || amazon-linux-extras install -y ansible2
 fi
